Refactor file organization to be consistent by command
