+++ cmd/fleetctl/generate_test.go
